Move appcontext test case to integration
