+++ tests/integration/integration_test_case.py
@@ -14,7 +14,7 @@ from app.keys import KEY_PURPOSE_AUTHENTICATION, KEY_PURPOSE_SUBMISSION
 from app.setup import create_app
 from app.utilities.json import json_loads
 from application import configure_logging
-from tests.app.app_context_test_case import MockDatastore
+from tests.integration.app_context_test_case import MockDatastore
 from tests.integration.create_token import TokenGenerator
 
 EQ_USER_AUTHENTICATION_RRM_PRIVATE_KEY_KID = "709eb42cfee5570058ce0711f730bfbb7d4c8ade"
