+++ tests/integration/routes/test_jwt_authentication.py
@@ -9,11 +9,11 @@ from sdc.crypto.encrypter import encrypt
 from sdc.crypto.key_store import KeyStore
 
 from app.keys import KEY_PURPOSE_AUTHENTICATION
-from tests.app.app_context_test_case import AppContextTestCase
 from tests.app.authentication import (
     TEST_DO_NOT_USE_SR_PUBLIC_KEY,
     TEST_DO_NOT_USE_UPSTREAM_PRIVATE_KEY,
 )
+from tests.integration.app_context_test_case import AppContextTestCase
 from tests.integration.integration_test_case import (
     EQ_USER_AUTHENTICATION_RRM_PRIVATE_KEY_KID,
     SR_USER_AUTHENTICATION_PUBLIC_KEY_KID,
